lesson3_hw

+++ src/main/java/lesson4_hw/service/HotelService.java
@@ -2,19 +2,27 @@ package lesson4_hw.service;
 
 import lesson4_hw.Exception.BadRequestException;
 import lesson4_hw.Exception.ObjectNotFoundInBDException;
+import lesson4_hw.Exception.PermissionException;
 import lesson4_hw.dao.HotelDAO;
+import lesson4_hw.dao.UserDAO;
 import lesson4_hw.model.Hotel;
+import lesson4_hw.model.User;
 
 public class HotelService {
     HotelDAO hotelDAO = new HotelDAO();
+    UserDAO userDAO = new UserDAO();
 
+    public Hotel save(Hotel hotel, User user) throws BadRequestException, PermissionException {
+        checkPermission(user);
 
-    public Hotel save(Hotel hotel) throws BadRequestException {
-        hotelDAO.checkIfHotelIsExist(hotel);
-        return hotelDAO.save(hotel);
+        if (hotelDAO.findHotelByName(hotel.getName()) == null) {
+            return hotelDAO.save(hotel);
+        }
+        throw new BadRequestException("Hotel with name = " + hotel.getName() + " is already registred.");
     }
 
-    public void delete(long id) {
+    public void delete(long id, User user) throws PermissionException {
+        checkPermission(user);
         try {
             hotelDAO.delete(id);
         } catch (ObjectNotFoundInBDException e) {
@@ -22,7 +30,8 @@ public class HotelService {
         }
     }
 
-    public Hotel update(Hotel hotel) {
+    public Hotel update(Hotel hotel,User user) throws PermissionException {
+        checkPermission(user);
         try {
             hotelDAO.update(hotel);
         } catch (ObjectNotFoundInBDException e) {
@@ -33,5 +42,10 @@ public class HotelService {
 
     public Hotel findById(long id) {
         return hotelDAO.findById(id);
+    }
+
+    private void checkPermission(User user) throws PermissionException {
+        if (user.getUserType().equals("USER") )
+            throw new PermissionException("User ID : " + user.getId() + "have not rights to make operation.");
     }
 }
