+++ src/main/java/lesson4_hw/service/RoomService.java
@@ -1,19 +1,23 @@
 package lesson4_hw.service;
 
 import lesson4_hw.Exception.ObjectNotFoundInBDException;
+import lesson4_hw.Exception.PermissionException;
 import lesson4_hw.dao.RoomDAO;
 import lesson4_hw.model.Room;
+import lesson4_hw.model.User;
 
 import java.util.List;
 
 public class RoomService {
     RoomDAO roomDAO = new RoomDAO();
-    public Room save(Room room) {
+    public Room save(Room room, User user) throws PermissionException {
+        checkPermission(user);
         RoomDAO.save(room);
         return room;
     }
 
-    public void delete(long id) {
+    public void delete(long id, User user) throws PermissionException {
+        checkPermission(user);
         try {
             roomDAO.delete(id);
         } catch (ObjectNotFoundInBDException e) {
@@ -21,7 +25,8 @@ public class RoomService {
         }
     }
 
-    public Room update(Room room) {
+    public Room update(Room room, User user) throws PermissionException {
+        checkPermission(user);
         try {
             roomDAO.update(room);
         } catch (ObjectNotFoundInBDException e) {
@@ -31,10 +36,15 @@ public class RoomService {
     }
 
     public Room findById(long id) {
-        return roomDAO.findById(id);
+        return roomDAO.finByID(id);
     }
 
     public List<Room> getAllRooms() {
         return roomDAO.getAllRooms();
+    }
+
+    private void checkPermission(User user) throws PermissionException {
+        if (user.getUserType().equals("USER"))
+            throw new PermissionException("User ID : " + user.getId() + "have not rights to make operation.");
     }
 }
